Change faraday version constraint to pessimistic
The current dependency constraint of '>=0.8.6' for faraday is causing
`bundle outdated` to error out on a project I am working on. This is
due to faraday-middleware being incompatible with faraday 0.9 and
bundler trying to update to faraday 0.9 from the lax balanced
constraint.

By switching to pessimistic versioning, we can avoid this dependency
problem without limiting users to any different version of faraday
since faraday-middleware shares this version constraint.